got rid of duplications
